feat(copilot): implement multi-agent router orchestration
