+++ Backend-RL/src/api/routers/legacy_routes.py
@@ -51,7 +51,7 @@ from models.domain import UploadedFile, TrainingRun, EvaluationResult
 from services import storage_service
 from services.queue_service import publish_training_job, ProgressListener, publish_ui_update
 from services.chat.chatbot import parse_demand_intent, action_to_human_message
-from services.chat.copilot import handle_copilot_message
+from services.chat.agents import handle_copilot_message
 
 # --- Add RL experiment modules to path (for demand.py, trainer.py, etc.) ---
 RL_DIR = os.path.join(os.path.dirname(__file__), "..", "experiments", "backend-implementation")
