+++ Backend-RL/src/services/chat/agents/__init__.py
@@ -1,0 +1,3 @@
+from .orchestrator import handle_copilot_message
+
+__all__ = ["handle_copilot_message"]
